+++ mobile/android/android-components/components/feature/sitepermissions/src/androidTest/java/mozilla/components/feature/sitepermissions/db/OnDeviceSitePermissionsStorageTest.kt
@@ -5,6 +5,7 @@
 package mozilla.components.feature.sitepermissions.db
 
 import android.content.Context
+import androidx.core.net.toUri
 import androidx.room.Room
 import androidx.room.testing.MigrationTestHelper
 import androidx.sqlite.db.framework.FrameworkSQLiteOpenHelperFactory
@@ -12,7 +13,6 @@ import androidx.test.core.app.ApplicationProvider
 import androidx.test.platform.app.InstrumentationRegistry
 import mozilla.components.feature.sitepermissions.SitePermissions
 import mozilla.components.feature.sitepermissions.SitePermissionsStorage
-import mozilla.components.support.ktx.kotlin.toUri
 import org.junit.After
 import org.junit.Assert.assertEquals
 import org.junit.Before