+++ mobile/android/android-components/components/support/ktx/src/test/java/mozilla/components/support/ktx/kotlin/StringTest.kt
@@ -84,6 +84,7 @@ class StringTest {
         assertTrue("Geo:1,-1".isGeoLocation())
     }
 
+    @Suppress("Deprecation")
     @Test
     fun toUri() {
         assertEquals(