+++ mobile/android/android-components/docs/changelog.md
@@ -32,6 +32,7 @@ permalink: /changelog/
   * Added `FetchDownloadManager`, an alternate download manager that uses a fetch `Client` instead of the native Android `DownloadManager`.
 
 * **support-ktx**
+  * Deprecated `String.toUri()` in favour of Android Core KTX.
   * Added `putCompoundDrawablesRelative` and `putCompoundDrawablesRelativeWithIntrinsicBounds`, aliases of `setCompoundDrawablesRelative` that use Kotlin named and default arguments.
 
 * **feature-prompts**
@@ -200,8 +201,8 @@ permalink: /changelog/
   * ⚠️ **This is a breaking change**: Added `url` parameter to `Session.Observer.onLoadRequest()`.
 
 * **support-ktx**
-  * Removed `Drawable.toBitmap()` in favour of the Android Core KTX version.
-  * Removed `Context.systemService()` in favour of the Android Core KTX version.
+  * ⚠️ **This is a breaking change**: Removed `Drawable.toBitmap()` in favour of the Android Core KTX version.
+  * ⚠️ **This is a breaking change**: Removed `Context.systemService()` in favour of the Android Core KTX version.
 
 * **browser-session**
   * Added `Session.hasParentSession` to indicate whether a `Session` was opened from a parent `Session` such as opening a new tab from a link context menu ("Open in new tab").